Moving CSV functions to a new place
