+++ mothra/tests/test_pipeline.py
@@ -13,7 +13,6 @@ from skimage.util import img_as_float
 
 
 PATH_TEST_FILES = 'mothra/tests/test_files'
-TEST_CSV_FILE = f'{PATH_TEST_FILES}/test_file.csv'
 TEST_INPUT_FILE = f'{PATH_TEST_FILES}/input_file.txt'
 TEST_INPUT_IMAGES = glob(f'{PATH_TEST_FILES}/test_input/*.JPG')
 TEST_TILTED_IMAGE =  f'{PATH_TEST_FILES}/test_input/BMNHE_1105737_17193_6eec94847b4939c6d117429d59829aac7a9fadf9.JPG'
@@ -83,31 +82,6 @@ def test_create_layout():
         assert ax is None
 
 
-def test_initialize_csv_file():
-    """Checks if CSV file is initialized properly, and if its filename
-    is correct.
-
-    Summary
-    -------
-    We pass a filename to pipeline.initialize_csv_file, check if the
-    file is created in disk, and if the name is the same as expected.
-
-    Expected
-    --------
-    result_csv is a file on disk, and its name is equal to expected_csv.
-    """
-    csv_fname = TEST_CSV_FILE
-    result_csv = pipeline.initialize_csv_file(csv_fname)
-
-    assert result_csv.is_file()
-
-    # deleting file, else next round of tests will have a different filename
-    result_csv.unlink()
-    expected_csv = Path(TEST_CSV_FILE)
-
-    assert expected_csv == result_csv
-
-
 def test_read_orientation():
     """Checks if orientation is extracted correctly from EXIF data.
 
@@ -124,29 +98,6 @@ def test_read_orientation():
     angle = pipeline.read_orientation(TEST_TILTED_IMAGE)
 
     assert angle == 90
-
-
-def test_check_aux_file():
-    """Checks if filename is updated correctly if file already exists
-    in disk.
-
-    Summary
-    -------
-    We provide the path of a file that exists in disk, and check
-    if result_fname was updated correctly (a number was added to
-    the end of the filename).
-
-    Expected
-    --------
-    "result.csv.test" already exists in disk; pipeline._check_aux_file
-    should return "result.csv_1.test".
-    """
-    expected_fname = Path('result.csv_1.test')
-
-    filename = Path(f'{PATH_TEST_FILES}/result.csv.test')
-    result_fname = pipeline._check_aux_file(filename)
-
-    assert result_fname == expected_fname
 
 
 def test_process_paths_in_input():
@@ -183,56 +134,6 @@ def test_read_paths_in_file():
     image_paths = pipeline._read_paths_in_file(TEST_INPUT_FILE)
 
     assert image_paths.sort() == TEST_INPUT_IMAGES.sort()
-
-
-def test_write_csv_data():
-    """ Check if data is written properly in the CSV file.
-
-    Summary
-    -------
-    We create a CSV file with test data and check if the
-    function pipeline._write_csv_data writes the data
-    properly to the file.
-
-    Expected
-    --------
-    Rows in expected_lines and csv_fname are equal.
-    """
-    csv_fname = f'{PATH_TEST_FILES}/test.csv'
-    image_name = 'test_image'
-    dist_mm = {
-        "dist_l": 'test_l',
-        "dist_r": 'test_r',
-        "dist_l_center": 'test_dcl',
-        "dist_r_center": 'test_dcr',
-        "dist_span": 'test_span',
-        "dist_shoulder": 'test_shoulder'
-    }
-    position = 'test_pos'
-    gender = 'test_gender'
-    probabilities = "['testprob_1', 'testprob_2', 'testprob_3']"
-
-    expected_lines = [
-        ['image_id', 'left_wing (mm)', 'right_wing (mm)',
-         'left_wing_center (mm)', 'right_wing_center (mm)',
-         'wing_span (mm)', 'wing_shoulder (mm)', 'position',
-         'gender', 'probabilities'],
-        ['test_image', 'test_l', 'test_r', 'test_dcl', 'test_dcr',
-        'test_span', 'test_shoulder', 'test_pos', 'test_gender',
-        "['testprob_1', 'testprob_2', 'testprob_3']"]
-        ]
-
-    result_csv = pipeline.initialize_csv_file(csv_fname)
-    with open(csv_fname, 'a') as csv:
-        pipeline._write_csv_data(csv, image_name, dist_mm, position, gender,
-                                 probabilities)
-
-    with open(csv_fname, newline='') as csv:
-        for idx, row in enumerate(reader(csv)):
-            assert expected_lines[idx] == row
-
-    # deleting file, else next round of tests will have more rows
-    result_csv.unlink()
 
 
 def test_read_filenames_in_folder():
